// src/app/shared/enums/errors.enum.ts
export enum ErrorsEnum {
	IsString = "IsString",
	IsNumber = "IsNumber",
	IsNotEmpty = "IsNotEmpty",
	IsEmail = "IsEmail",
	IsObject = "IsObject",
	IsDate = "IsDate",
	IsArray = "IsArray",
	MaxLength = "MaxLength",
	MinLength = "MinLength",
	Max = "Max",
	Min = "Min",
	InvalidEnum = "InvalidEnum",
	InvalidEncryptionLength = "InvalidEncryptionLength",
	InvalidPassword = "InvalidPassword",
	InvalidEncryption = "InvalidEncryption",
	UserAlreadyExist = "UserAlreadyExist",
	UserNotExist = "UserNotExist",
	UserNotVerified = "UserNotVerified",
	InvalidVerificationCode = "InvalidVerificationCode",
	AuthorizationHeaderIsMissing = "AuthorizationHeaderIsMissing",
	TokenIsMissing = "TokenIsMissing",
	InvalidToken = "InvalidToken",
	UserNotFound = "UserNotFound",
	Forbidden = "Forbidden",
	NotFound = "NotFound",
	InternalServerError = "InternalServerError"
}