// src/app/shared/enums/events.enum.ts
export enum EventsEnum {
	SOLANA_MESSAGE = "SOLANA_MESSAGE",
	SOLANA_TRANSACTION = "SOLANA_TRANSACTION",
	SOLANA_TRANSACTION_FAILED = "SOLANA_TRANSACTION_FAILED",
	HELIUS_OPEN = "HELIUS_OPEN",
	HELIUS_ERROR = "HELIUS_ERROR",
	HELIUS_CLOSE = "HELIUS_CLOSE",
	POOL_CREATED = "POOL_CREATED",
	POOL_UPDATED = "POOL_UPDATED",
	POOL_DELETED = "POOL_DELETED",
	TRADING_CREATED = "TRADING_CREATED",
	TRADING_UPDATED = "TRADING_UPDATED",
	TRADING_DELETED = "TRADING_DELETED",
	STRATEGY_CREATED = "STRATEGY_CREATED",
	STRATEGY_UPDATED = "STRATEGY_UPDATED",
	STRATEGY_DELETED = "STRATEGY_DELETED",
	STRATEGIES_CREATED = "STRATEGIES_CREATED",
	TRADING_TOKEN_CREATED = "TRADING_TOKEN_CREATED",
	TRADING_TOKEN_UPDATED = "TRADING_TOKEN_UPDATED",
	TRADING_TOKEN_DELETED = "TRADING_TOKEN_DELETED",
	TRADING_TOKENS_UPDATED = "TRADING_TOKENS_UPDATED",
	TRADING_TRANSACTION = "TRADING_TRANSACTION",
	NEW_POOL_DETECTED = "NEW_POOL_DETECTED",
	MILESTONE_CREATED = "MILESTONE_CREATED",
	MILESTONE_UPDATED = "MILESTONE_UPDATED",
	MILESTONE_DELETED = "MILESTONE_DELETED",
	MILESTONES_CREATED = "MILESTONES_CREATED",
	MILESTONE_CONFIRMED = "MILESTONE_CONFIRMED",
	MILESTONE_CHECKED = "MILESTONE_CHECKED",
	CONDITIONS_CREATED = "CONDITIONS_CREATED",
	CONDITIONS_GROUP_CREATED = "CONDITIONS_GROUP_CREATED",
	CONDITIONS_GROUPS_CREATED = "CONDITIONS_GROUPS_CREATED",
	CONDITIONS_GROUP_UPDATED = "CONDITIONS_GROUP_UPDATED"
}